+++ src/streamlabs/default-settings.ts
@@ -1,4 +1,5 @@
 export default {
   showPronouns: true,
   showFrogEmotes: true,
+  capitalizePronouns: true,
 };